refactor: ♻️ use svelte action for modal background clicks
